Allow conversion of format without processing the date
There are cases when the format is coming from PHP and we don't need the date formatted but we want to convert the format string from PHP style to Moment.js. This patch is simply externalizing the replacement of the format string. Now, in your app, you can do something like:

```
var formatFromPhp = 'Y-m-d';

// This will output "YYYY-MM-DD".
console.log(moment().phpConvertFormat(formatFromPhp));
```